Add offwork command to calculate end of work hours
Implemented the 'offwork' command in 'offwork.go'. This new command calculates the standard off-work time as well as one and two hour overtime periods based on the provided on-work time. By including this feature, the application can now assist in time keeping and aid employees in maintaining their working hours.